+++ rpi/semmVis/playboardscene.h
@@ -35,7 +35,7 @@ public slots:
     // add meepes of player color with idtags (uint16_t) on fieldIds (uint8_t)
     void addMeeples( uint8_t color, std::map< uint16_t, uint8_t > tagsFields );
 
-    void setMeeplePos( uint8_t color, uint8_t tag, uint8_t toFieldId );
+    void setMeeplePos( uint8_t color, uint16_t tag, uint8_t toFieldId );
 
 private:
